+++ test/Unit/TRegx/CleanRegex/Match/MatchPattern/groups/only/onlyNegative/MatchPatternTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace Test\Unit\TRegx\CleanRegex\Match\MatchPattern\groups\all;
+namespace Test\Unit\TRegx\CleanRegex\Match\MatchPattern\groups\only\onlyNegative;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
@@ -12,50 +12,49 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_groups()
+    public function shouldThrow_forMatching()
     {
         // given
         $pattern = new MatchPattern(new Pattern('(?<two>[A-Z][a-z])?(?<rest>[a-z]+)'), 'Nice Matching Pattern');
 
-        // when
-        $twoGroups = $pattern->group('two')->all();
-        $restGroups = $pattern->group('rest')->all();
-
         // then
-        $this->assertEquals(['Ni', 'Ma', 'Pa'], $twoGroups);
-        $this->assertEquals(['ce', 'tching', 'ttern'], $restGroups);
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Negative limit -2');
+
+        // when
+        $pattern->group('two')->only(-2);
     }
 
     /**
      * @test
      */
-    public function shouldGet_unmatchedGroups()
+    public function shouldThrow_unmatchedGroups()
     {
         // given
         $pattern = new MatchPattern(new Pattern('(?<hour>\d\d)?:(?<minute>\d\d)?'), 'First->11:__   Second->__:12   Third->13:32');
 
-        // when
-        $hours = $pattern->group('hour')->all();
-        $minutes = $pattern->group('minute')->all();
-
         // then
-        $this->assertEquals(['11', null, '13'], $hours);
-        $this->assertEquals([null, '12', '32'], $minutes);
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Negative limit -2');
+
+        // when
+        $pattern->group('hour')->only(-2);
     }
 
     /**
      * @test
      */
-    public function shouldReturnEmptyArray_onNotMatchedSubject()
+    public function shouldThrow_onNotMatchedSubject()
     {
         // given
         $pattern = new MatchPattern(new Pattern('(?<two>[A-Z][a-z])?(?<rest>[a-z]+)'), 'NOT MATCHING');
 
-        // when
-        $groups = $pattern->group('two')->all();
-
         // then
-        $this->assertEquals([], $groups);
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Negative limit -2');
+
+        // when
+        $pattern->group('two')->only(-2);
     }
 
     /**
@@ -71,7 +70,7 @@ class MatchPatternTest extends TestCase
         $this->expectExceptionMessage("Nonexistent group: 'missing'");
 
         // when
-        $pattern->group('missing')->all();
+        $pattern->group('missing')->only(-2);
     }
 
     /**
@@ -87,7 +86,7 @@ class MatchPatternTest extends TestCase
         $this->expectExceptionMessage("Nonexistent group: 'missing'");
 
         // when
-        $pattern->group('missing')->all();
+        $pattern->group('missing')->only(-2);
     }
 
     /**
@@ -103,6 +102,6 @@ class MatchPatternTest extends TestCase
         $this->expectExceptionMessage("Group name must be an alphanumeric string sequence starting with a letter, or an integer");
 
         // when
-        $pattern->group('2invalid')->all();
+        $pattern->group('2invalid')->only(-2);
     }
 }